+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add UI selection persistence between sessions [#512](https://github.com/nbbrd/beanzooka/issues/512)
 - Add diff bar overlay in list editor [#513](https://github.com/nbbrd/beanzooka/issues/513)
+- Add search on path for autofill JDK [#395](https://github.com/nbbrd/beanzooka/issues/395)
 
 ### Changed